requirements.txt
django
djangorestframework
drf-yasg
psycopg2-binary
pillow
python-decouple
python-slugify
djangorestframework-simplejwt
celery
redis
